Move catch struct into game
